Тест добавления шмотки WIP
